ui: Remove buttons entirely from track shell when not hovered
Currently, some track shell buttons are hidden when the track is not
hovered in order to avoid cluttering the UI visually, however for very
reasonable reasons they still take space out of the shell which is
otherwise used for the title.

When titles are very long it would be nice if we could use that space
for the title instead of the buttons, especially if the track has a
chip on it too (e.g. main thread).

This patch makes it so that the buttons are actually removed (display:
none) when the track shell is not hovered, so the space may be used
for the title.

Change-Id: I50fac02972f5e6af000bcd6620b37f9e9ad4b10e