ui: Remove ExplorePage's dependency on DataGrid's filter model (#3386)
Looks like `FilterDefinition` was being used through the explore page
which made it difficult to add an 'in' filter to without also adding
support for it in the explore page too.

This PR:
- Extracts the `FilterDefinition` from explore page, replacing it with
the already defined `UIFilter` instead.
- Change the DataGrid API from `onFilterChange(newFilterState)` to a
pair of functions `onFilterAdd(newFilter)` and `onFilterRemove(index)`
which allows a lot more flexibility in where and how the filters are
stored by the parent. In short it means that the parent doesn't have to
support all the native filter types or the parent could support more
filter types than those that are supported in DataGrid, and instead can
just pick and choose the filters it wants to use. This essentiallly
decouples the implementation of DataGrid filters from its parents.
- Rename `FilterDefinition` to `DataGridFilter` to make it clear it's
part of DataGrid.
- Never pass the filters to the datagrid in the explore page, as:
- they are already displayed on the selected node, so don't need to be
displayed twice.
- passing them to datagrid also means that it applies those filters when
reading from the datasource which the explore page has already
pre-filtered, so there's no point in filtering twice.